Put generators in main schema
