+++ tests/toolbox/sample_loader/test_generate_sample.py
@@ -1,4 +1,5 @@
 from toolbox.sample_loader.generate_sample import SampleGenerator
+from toolbox.sample_loader.schema import SCHEMA
 from toolbox.sample_loader.validate_sample import SampleValidator
 
 
@@ -7,7 +8,7 @@ def test_generated_sample_passes_validation(tmp_path):
     generated_sample_path = tmp_path.joinpath('generated_sample.csv')
 
     # When
-    SampleGenerator().generate_sample_file(20, generated_sample_path)
+    SampleGenerator(SCHEMA).generate_sample_file(20, generated_sample_path)
 
     # Then
     validation_failures = SampleValidator().validate(
